redirect stderr to dedicated buffer
Because `:stderr` is not specified in the `make-process` block, `stdout` and
`stderr` are intermingled. `tex2svg` can raise warnings (e.g., a deprecation
warning regarding `punycode` that seems to affect many of its dependencies), and
these warnings are sent to `stderr`. If the warning is comingled with stdout,
then what we get is no longer valid svg.

This commit adds support for `org-latex-impatient--stderr-buffer` to parallel
support for `org-latex-impatient--output-buffer` and then uses it as a place to
redirect stderr