default to finding tex2svg on user's $PATH
If `tex2svg` is installed in such a way that it is available on `$PATH` (from
within emacs), then there's no need for the user to install it manually.

I've set the default value of `org-latex-impatient-tex2svg-bin` to simply
`tex2svg` (as opposed to the empty string which it was previously set to); this
default will work if emacs can find `tex2svg` on the path. I've also updated the
documentation to specify that the user *may* need to set the value (since
otherwise things may just work). I've also adjusted the documentation to suggest
a global installation of `mathjax-node-cli` (and fixed some indentation while I
was at it), since IMHO this makes it more likely that `tex2svg` will be on
`$PATH` (e.g., on MacOS, if `node` is installed using `homebrew`, then `npm
install -g mathjax-node-cli` will put `tex2svg` under
`/opt/homebrew/bin/tex2svg`, which is likely to be on `$PATH`