ensure calls to equal-or-member function use org-latex-impatient
org-latex-impatient provides a useful function called
`org-latex-impatient--equal-or-member`, which it calls in a few places. however,
in some places (mostly in code handling markdown mode), it instead calls
`yang/equal-or-member`. I assume this is an artifact of the development cycle
and that the latter is a function available to the developer, but not in the
package. I've simply renamed any instances of `yang/equal-or-member` to
`org-latex-impatient--equal-or-member`.